docs(widgets): fold missing-data illustrations into the charting principles doc (LFE-10694)
Move the gap-bearing missing-data stories (gap vs zero, honest vs
bridged, isolated dots, staggered gaps, axis-kept empty buckets, area
variants) into the Illustrations stories and weave them into section 2
of the Principles article, so the doc shows how production renders
missing data. Drop the no-gap sanity/playground stories and the
standalone MissingData stories file.
