+++ web/src/features/widgets/chart-library/ChartingPrinciples.stories.tsx
@@ -7,6 +7,7 @@ import {
   XAxis,
   YAxis,
 } from "recharts";
+import { AreaChartTimeSeries } from "./AreaChartTimeSeries";
 import { LineChartTimeSeries } from "./LineChartTimeSeries";
 import { formatMetric } from "./utils";
 import { type DataPoint, type MetricFormatterFunction } from "./chart-props";
@@ -33,15 +34,30 @@ const sparse = [
 const ExampleCard = ({
   label,
   caption,
+  tall = false,
   children,
 }: {
   label: string;
   caption: string;
+  /** Production-component demos need room for legends + axis labels. */
+  tall?: boolean;
   children: React.ReactNode;
 }) => (
-  <div className="flex min-w-[210px] flex-1 flex-col gap-1">
+  <div
+    className={
+      tall
+        ? "flex min-w-[340px] flex-1 flex-col gap-1"
+        : "flex min-w-[210px] flex-1 flex-col gap-1"
+    }
+  >
     <div className="text-xs font-bold tracking-wide uppercase">{label}</div>
-    <div className="bg-background h-[150px] w-full rounded-md border p-2">
+    <div
+      className={
+        tall
+          ? "bg-background h-[240px] w-full rounded-md border p-2"
+          : "bg-background h-[150px] w-full rounded-md border p-2"
+      }
+    >
       {children}
     </div>
     <div className="text-muted-foreground w-full text-xs">{caption}</div>
@@ -163,7 +179,7 @@ export const NullHandling = meta.story({
       </ExampleCard>
       <ExampleCard
         label="Zero-fill ✗"
-        caption="Invents a measurement of 0. A lie, unless stacking demands it."
+        caption="Invents a measurement of 0. A lie for anything non-additive; only counts/sums may fill zeros."
       >
         <ResponsiveContainer>
           <LineChart data={zeroFilled} margin={{ top: 8, right: 8, bottom: 0 }}>
@@ -336,5 +352,261 @@ export const BoundTheFrame = meta.story({
         metricFormatter={msFormatter}
       />
     </ChartFrame>
+  ),
+});
+
+// ── V2 in production · missing data through the real components ─────────────
+// These mirror the exact data shape the widget transforms hand the charts
+// (LFE-10694): a series' missing bucket is an ABSENT DataPoint, and a bucket
+// where nothing measured anything arrives as a dimension-less bucket MARKER
+// (`dimension: undefined, metric: null`) — what the query's gap-fill
+// placeholder rows become.
+
+type SeriesSpec = {
+  name: string;
+  /** One entry per bucket; `null` = this series measured nothing there. */
+  values: (number | null)[];
+};
+
+/** Buckets are day labels 7/1..7/N; markers keep all-empty buckets on the axis. */
+const buildGappedData = (series: SeriesSpec[]): DataPoint[] => {
+  const bucketCount = Math.max(...series.map((s) => s.values.length));
+  const points: DataPoint[] = [];
+  for (let bucket = 0; bucket < bucketCount; bucket++) {
+    const label = `7/${bucket + 1}`;
+    let bucketHasData = false;
+    for (const { name, values } of series) {
+      const value = values[bucket];
+      if (value != null) {
+        bucketHasData = true;
+        points.push({ time_dimension: label, dimension: name, metric: value });
+      }
+    }
+    if (!bucketHasData) {
+      points.push({
+        time_dimension: label,
+        dimension: undefined,
+        metric: null,
+      });
+    }
+  }
+  return points;
+};
+
+// One series, healthy on both sides of a 3-bucket hole (the classic lie shape).
+const midRangeGap: SeriesSpec[] = [
+  {
+    name: "gpt-4o",
+    values: [820, 940, 760, null, null, null, 880, 810, 990, 870],
+  },
+];
+
+// Multi-series where the gaps don't line up — every bucket has SOME data, so
+// only per-series cells are missing (no markers involved).
+const staggeredGapSeries: SeriesSpec[] = [
+  {
+    name: "gpt-4o",
+    values: [820, 940, 760, null, null, 830, 880, 810, 990, 870],
+  },
+  {
+    name: "claude-3-5-sonnet",
+    values: [1450, null, null, 1380, 1520, 1490, null, null, null, 1400],
+  },
+  {
+    name: "gpt-4o-mini",
+    values: [310, 290, 340, 300, null, null, null, 280, 320, 330],
+  },
+  {
+    name: "mistral-7b",
+    values: [null, null, 540, 560, 520, null, 580, 610, null, null],
+  },
+];
+
+// Isolated points: a lone first bucket, a mid-gap single, a lone last bucket,
+// and a series that is a segment plus one stranded value.
+const isolatedPointSeries: SeriesSpec[] = [
+  {
+    name: "first-bucket-only",
+    values: [720, null, null, null, null, null, null, null, null, null],
+  },
+  {
+    name: "single-mid-gap",
+    values: [null, null, null, null, 1350, null, null, null, null, null],
+  },
+  {
+    name: "last-bucket-only",
+    values: [null, null, null, null, null, null, null, null, null, 980],
+  },
+  {
+    name: "segment-plus-stranded",
+    values: [420, 460, 440, null, null, null, null, 510, null, null],
+  },
+];
+
+// Long empty stretch (buckets exist only as markers) between two active eras.
+const twoEras: SeriesSpec[] = [
+  {
+    name: "gpt-4o",
+    values: [88, 96, 74, null, null, null, null, null, 91, 83],
+  },
+  {
+    name: "claude-3-5-sonnet",
+    values: [54, 61, null, null, null, null, null, null, 58, 66],
+  },
+];
+
+export const GapVersusZero = meta.story({
+  render: () => (
+    <div className="flex flex-wrap gap-5 p-1">
+      <ExampleCard
+        tall
+        label='missingValue="gap" (avg / percentiles)'
+        caption="No latency exists on a bucket without generations — the line breaks. Fabricating a 0 would deflate the chart."
+      >
+        <LineChartTimeSeries
+          data={buildGappedData(midRangeGap)}
+          missingValue="gap"
+          metricFormatter={msFormatter}
+        />
+      </ExampleCard>
+      <ExampleCard
+        tall
+        label='missingValue="zero" (count / sum)'
+        caption="Zero events on a bucket without data is the truth — the same shape stays continuous, touching 0."
+      >
+        <LineChartTimeSeries
+          data={buildGappedData(midRangeGap)}
+          missingValue="zero"
+        />
+      </ExampleCard>
+    </div>
+  ),
+});
+
+export const HonestVersusBridged = meta.story({
+  render: () => (
+    <div className="flex flex-wrap gap-5 p-1">
+      <ExampleCard
+        tall
+        label="default (honest)"
+        caption="Gap semantics: three empty buckets stay on the axis and the line breaks over them."
+      >
+        <LineChartTimeSeries
+          data={buildGappedData(midRangeGap)}
+          missingValue="gap"
+          metricFormatter={msFormatter}
+        />
+      </ExampleCard>
+      <ExampleCard
+        tall
+        label="connectNulls (the old default, now opt-in)"
+        caption="The bridge draws values that were never measured. Reserved for series that semantically continue across a gap."
+      >
+        <LineChartTimeSeries
+          data={buildGappedData(midRangeGap)}
+          missingValue="gap"
+          connectNulls
+          metricFormatter={msFormatter}
+        />
+      </ExampleCard>
+    </div>
+  ),
+});
+
+export const IsolatedPoints = meta.story({
+  render: () => (
+    <div className="flex flex-wrap gap-5 p-1">
+      <ExampleCard
+        tall
+        label="neighborless values"
+        caption="A value with gaps on both sides spans no line segment — the dot is what keeps it visible (first bucket, mid-gap, last bucket, and a stranded point after a segment)."
+      >
+        <LineChartTimeSeries
+          data={buildGappedData(isolatedPointSeries)}
+          missingValue="gap"
+          legendPosition="above"
+          metricFormatter={msFormatter}
+        />
+      </ExampleCard>
+    </div>
+  ),
+});
+
+export const StaggeredSeriesGaps = meta.story({
+  render: () => (
+    <div className="flex flex-wrap gap-5 p-1">
+      <ExampleCard
+        tall
+        label="per-series gaps, every bucket has some data"
+        caption="Each series breaks over its own missing buckets while the others keep drawing — no series borrows another's shape."
+      >
+        <LineChartTimeSeries
+          data={buildGappedData(staggeredGapSeries)}
+          missingValue="gap"
+          legendPosition="above"
+          metricFormatter={msFormatter}
+        />
+      </ExampleCard>
+    </div>
+  ),
+});
+
+export const EmptyBucketsStayOnAxis = meta.story({
+  render: () => (
+    <div className="flex flex-wrap gap-5 p-1">
+      <ExampleCard
+        tall
+        label="gap semantics"
+        caption="Buckets 7/4–7/8 exist only as markers. The axis keeps them, so the empty era is visible instead of time compressing away."
+      >
+        <LineChartTimeSeries
+          data={buildGappedData(twoEras)}
+          missingValue="gap"
+          legendPosition="above"
+        />
+      </ExampleCard>
+      <ExampleCard
+        tall
+        label="zero semantics"
+        caption="Same data as counts: the empty era reads as an honest flat 0, not as absence."
+      >
+        <LineChartTimeSeries
+          data={buildGappedData(twoEras)}
+          missingValue="zero"
+          legendPosition="above"
+          legendSummary="sum"
+        />
+      </ExampleCard>
+    </div>
+  ),
+});
+
+export const AreaGaps = meta.story({
+  render: () => (
+    <div className="flex flex-wrap gap-5 p-1">
+      <ExampleCard
+        tall
+        label='area · missingValue="gap"'
+        caption="The fill stops where the data stops; isolated values still get their dot."
+      >
+        <AreaChartTimeSeries
+          data={buildGappedData(staggeredGapSeries)}
+          missingValue="gap"
+          subtleFill
+          metricFormatter={msFormatter}
+        />
+      </ExampleCard>
+      <ExampleCard
+        tall
+        label='area · missingValue="zero"'
+        caption="Additive semantics: the fill dips to 0 across empty buckets."
+      >
+        <AreaChartTimeSeries
+          data={buildGappedData(twoEras)}
+          missingValue="zero"
+          subtleFill
+        />
+      </ExampleCard>
+    </div>
   ),
 });
